Make QIcon generate high-dpi pixmaps.
QIcon::pixmap() can now return pixmaps that are larger
than the requested size. This is a behavior change
that may require updating existing application code.

Add new pixmap() and actualSize() overloads that take
a window pointer. Since these are new functions they
can potentially always return "large" pixmaps. The
window pointer allows QIcon to return a pixmap suitable
for the display the window is on.

Update the existing pixmap() and actualSize() functions
to return high-dpi pixmaps when Qt::AA_UseHighDPIIamges
is set. This is implemented by calling the new functions
with a null QWindow pointer.

Change-Id: I2eefc4ed4aa48b036a7019446694e56213070efa